labs no1 report

+++ dbpu-l1/zad4/zad4.1.cpp
@@ -6,7 +6,7 @@ int main(){
 	long long abc[10];
 	long long def[10];
 
-	for(long long i=0;i<30;i++) abc[i]=i;
+	for(int i=0;i<30;i++) abc[i]=i;
 	abc[0]++;
-	for(long long i=0;i<10;i++) cout << def[i] << endl;
+	for(int i=0;i<10;i++) cout << def[i] << endl;
 }
